applied name router permission to update view step 2

+++ lib/app/modules/authentication/views/admin/permissions/widgets/permission_card.dart
@@ -10,6 +10,8 @@ import 'package:redstar_hightech_backend/app/modules/product/controllers/product
 import 'package:redstar_hightech_backend/app/services/database_service.dart';
 import 'package:safe_url_check/safe_url_check.dart';
 
+import '../../../../../middleware/auth_middleware.dart';
+import '../../../../controllers/authentication_controller.dart';
 import '../../../../controllers/permission_controller.dart';
 import '../../../../controllers/permission_controller.dart';
 import '../../../../models/permission_model.dart';
@@ -34,7 +36,26 @@ class PermissionCard extends StatelessWidget {
 
   Future<void> _onDeleteData(
       BuildContext context, Permission permission) async {
-    permissionController.deletePermission(permission);
+    if (AuthorizationMiddleware.checkPermission(
+        Get.find<AuthenticationController>(),
+        Get.find<UserController>(),
+        "/product/delete")) {
+      print("Check Delete route permission valid");
+      permissionController.deletePermission(permission);
+      //   Navigator.of(context).pop();
+    }
+    {
+      Get.snackbar(
+          "Delete product", "You don't have permission to delete product",
+          icon: const Icon(Icons.warning_amber),
+          margin: const EdgeInsets.only(left: 10, right: 10, bottom: 20),
+          backgroundColor: Colors.red,
+          snackPosition: SnackPosition.BOTTOM);
+      print(
+        "Check Delete route permission not valid",
+      );
+    }
+
     //   Navigator.of(context).pop();
   }
