+++ lib/app/modules/authentication/widgets/user_widget.dart
@@ -14,6 +14,9 @@ import '../../../constants/app_theme.dart';
 import '../../../routes/app_pages.dart';
 import 'package:intl/intl.dart';
 
+import '../../middleware/auth_middleware.dart';
+import '../controllers/authentication_controller.dart';
+
 enum Options { Edit, Delete }
 
 class UserCard extends StatelessWidget {
@@ -30,7 +33,26 @@ class UserCard extends StatelessWidget {
       : super(key: key);
 
   Future<void> _onDeleteData(BuildContext context, User user) async {
-    userController.deleteUser(user);
+    if (AuthorizationMiddleware.checkPermission(
+        Get.find<AuthenticationController>(),
+        Get.find<UserController>(),
+        "/product/delete")) {
+      print("Check Delete route permission valid");
+      userController.deleteUser(user);
+      //   Navigator.of(context).pop();
+    }
+    {
+      Get.snackbar(
+          "Delete product", "You don't have permission to delete product",
+          icon: const Icon(Icons.warning_amber),
+          margin: const EdgeInsets.only(left: 10, right: 10, bottom: 20),
+          backgroundColor: Colors.red,
+          snackPosition: SnackPosition.BOTTOM);
+      print(
+        "Check Delete route permission not valid",
+      );
+    }
+
     //   Navigator.of(context).pop();
   }
