+++ lib/app/modules/category/widgets/category_widget.dart
@@ -14,6 +14,8 @@ import '../../../constants/app_theme.dart';
 import '../../../routes/app_pages.dart';
 import 'package:intl/intl.dart';
 
+import '../../authentication/controllers/authentication_controller.dart';
+import '../../middleware/auth_middleware.dart';
 import '../../product/models/product_model.dart';
 import '../controllers/category_controller.dart';
 
@@ -35,7 +37,26 @@ class CategoryCard extends StatelessWidget {
       : super(key: key);
 
   Future<void> _onDeleteData(BuildContext context, Category category) async {
-    categoryController.deleteCategory(category);
+    if (AuthorizationMiddleware.checkPermission(
+        Get.find<AuthenticationController>(),
+        Get.find<UserController>(),
+        "/product/delete")) {
+      print("Check Delete route permission valid");
+      categoryController.deleteCategory(category);
+      //   Navigator.of(context).pop();
+    }
+    {
+      Get.snackbar(
+          "Delete product", "You don't have permission to delete product",
+          icon: const Icon(Icons.warning_amber),
+          margin: const EdgeInsets.only(left: 10, right: 10, bottom: 20),
+          backgroundColor: Colors.red,
+          snackPosition: SnackPosition.BOTTOM);
+      print(
+        "Check Delete route permission not valid",
+      );
+    }
+
     //   Navigator.of(context).pop();
   }
